Add SSLMODE environment variable, and connect retry for postgres driver
Adds the usql environment variable `SSLMODE`, read by default from
`$ENV{USQL_SSLMODE}` or `$ENV{SSLMODE}`. When not set by the OS's
environment, the default value will be set to `retry`.

Adds functionality to the `postgres` driver to reconnect when usql's
`SSLMODE=="retry"`. Attempts to reconnect by adding `sslmode=disable` to
the connection URL. The driver will only attempt to reconnect when no
`?sslmode=` parameter is present on the connection URL, `SSLMODE==retry`
and when the "SSL is not enabled on the server" error is encountered
during a database ping on opening the connection. This behavior better
mimics the `psql` command-line.