fix: strip trailing semicolon in cosmosdb and unify implementation
cosmosdb, like several other databases, needs trailing semicolons to be
removed from the statement before it is passed to the driver.

This fix applies that to cosmosdb and makes the other databases with identical
processing use the same implementation - presto, trino, athena, SAP ASE.

Drivers that strip trailing semicolon as part of other processing are
not included in the refactoring to keep scope small.

Testing: will run usql with trino, and cosmosdb