Replace unmaintained gohxs/readline with latest chzyer/readline
The Go library `gohxs/readline` has no commits for past 8 years. It was
originally forked from `chzyer/readline`, which continues to be
maintained, so switch back to it to ensure continued maintenance and
updates.

This change necessitated adapting the output filtering mechanism due to
API incompatibilities that manifested as errors:

    rline/rline.go:128:16: l.Inst.Config.Output undefined (type *readline.Config has no field or method Output)

    rline/rline.go:130:7: cfg.FuncFilterOutput undefined (type *readline.Config has no field or method FuncFilterOutput)

Specifically, the `readline.Config` no longer provides direct output
fields or filter functions. To address this, implement a `filterWriter`
to wrap the standard output stream, ensuring existing output filtering
functionality is preserved.

Closes: #528